+++ packages/ng/demo/app/select/searcher-select/searcher-select-demo.intl.ts
@@ -1,0 +1,9 @@
+import {LuSelectSearchIntl} from '../../../../src/app/select/searcher/';
+
+import {Injectable} from '@angular/core';
+
+export function getOverrideLuSelectSearchIntl(): LuSelectSearchIntl {
+	const luSelectSearchIntl = new LuSelectSearchIntl();
+	luSelectSearchIntl.noResultsLabel = 'Aucun résultat';
+	return luSelectSearchIntl;
+}