+++ packages/ng/src/app/select/option/feeder/index.ts
@@ -1,0 +1,2 @@
+export {LuOptionFeederDirective} from './select-option-feeder.directive';
+export {ASelectOptionFeeder, ISelectOptionFeeder} from './select-option-feeder.model';