+++ packages/ng/src/app/select/option/feeder/select-option-feeder.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { LuOptionFeederDirective } from './select-option-feeder.directive';
+
+@NgModule({
+	imports: [
+		CommonModule,
+	],
+	declarations: [
+		LuOptionFeederDirective,
+	],
+	exports: [
+		LuOptionFeederDirective,
+	]
+})
+export class LuSelectOptionFeederModule { }
+