+++ packages/ng/src/app/select/option/select-option.component.spec.ts
@@ -130,6 +130,7 @@ describe('LuSelectOption', () => {
 			});
 			const fixture = TestBed.createComponent(WrapperOptionElementComponent);
 			const wrapper = fixture.componentInstance;
+			fixture.detectChanges();
 			const option: LuSelectOption<any> = fixture.debugElement.query(By.directive(LuSelectOption)).componentInstance;
 
 			// Act
@@ -145,6 +146,7 @@ describe('LuSelectOption', () => {
 			});
 			const fixture = TestBed.createComponent(WrapperOptionElementComponent);
 			const wrapper = fixture.componentInstance;
+			fixture.detectChanges();
 			const option: LuSelectOption<any> = fixture.debugElement.query(By.directive(LuSelectOption)).componentInstance;
 
 			// Act
@@ -177,6 +179,7 @@ describe('LuSelectOption', () => {
 			});
 			const fixture = TestBed.createComponent(LuSelectOption);
 			const option = fixture.componentInstance;
+			fixture.detectChanges();
 
 			// Act
 			// Set the model
@@ -201,6 +204,7 @@ describe('LuSelectOption', () => {
 			});
 			const fixture = TestBed.createComponent(LuSelectOption);
 			const option = fixture.componentInstance;
+			fixture.detectChanges();
 			let liClickedCalled = false;
 			const event = new LuSelectOptionSelectionChange(option, true);
 			// Set the model
@@ -223,6 +227,36 @@ describe('LuSelectOption', () => {
 
 			// Assert
 			expect(liClickedCalled).toBeTruthy('You may have forgot the click handler on the `li` element');
+		});
+
+		it('It should be displayable at the initialisation', () => {
+			// Arrange
+			TestBed.configureTestingModule({
+				declarations: [LuSelectOption]
+			});
+			const fixture = TestBed.createComponent(LuSelectOption);
+			const option = fixture.componentInstance;
+
+			// Act
+
+			// Assert
+			expect(option.displayed).toBe(true);
+		});
+
+		it('It should be not display', () => {
+			// Arrange
+			TestBed.configureTestingModule({
+				declarations: [LuSelectOption]
+			});
+			const fixture = TestBed.createComponent(LuSelectOption);
+			const option = fixture.componentInstance;
+
+			// Act
+			option.displayed = false;
+			fixture.detectChanges();
+
+			// Assert
+			expect(option.displayed).toBe(false);
 		});
 
 });