+++ packages/ng/src/app/select/searcher/select-searcher.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { LuSelectSearcherComponent } from './select-searcher.component';
+import { LU_SELECT_SEARCH_INTL_PROVIDER } from './select-searcher.intl';
+
+@NgModule({
+	imports: [
+		CommonModule,
+		FormsModule,
+		BrowserModule,
+	],
+	declarations: [
+		LuSelectSearcherComponent,
+	],
+	exports: [
+		LuSelectSearcherComponent,
+	],
+	providers: [LU_SELECT_SEARCH_INTL_PROVIDER]
+})
+export class LuSelectSearcherModule { }
+