+++ packages/ng/src/app/select/select.module.ts
@@ -1,9 +1,12 @@
 import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { OverlayModule } from '@angular/cdk/overlay';
 import { LuSelectPickerModule } from './picker';
 import { LuSelectOptionModule } from './option';
 import { LuSelectClearerModule } from './clearer';
+import { LuSelectSearcherModule } from './searcher';
 import { LuSelectDirectiveModule } from './directive';
 import { LuSelect } from './select.component';
 import { LuEmptyModule } from '../empty/empty.module';
@@ -12,6 +15,8 @@ import { LuPopoverModule } from '../popover/popover.module';
 @NgModule({
 	imports: [
 		CommonModule,
+		FormsModule,
+		BrowserModule,
 		OverlayModule,
 		LuPopoverModule,
 		LuEmptyModule,
@@ -19,6 +24,7 @@ import { LuPopoverModule } from '../popover/popover.module';
 		LuSelectOptionModule,
 		LuSelectClearerModule,
 		LuSelectDirectiveModule,
+		LuSelectSearcherModule,
 	],
 	declarations: [
 		LuSelect,
@@ -28,6 +34,7 @@ import { LuPopoverModule } from '../popover/popover.module';
 		LuSelectOptionModule,
 		LuSelectClearerModule,
 		LuSelectDirectiveModule,
+		LuSelectSearcherModule,
 		LuSelect,
 	]
 })