+++ packages/ng/CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## In Dev
 ### New features
-
+- `lu-select-searcher` : Add a component for searching in a select
+- `IOptionFeeder` : An interface to implement to enable component that will manage options for a select
+- `ASelectOptionFeeder` : An abstract implentation of `IOptionFeeder`
 ### Fixes
 ### Breaking changes
+- `lu-select-picker` : `luOptions$` is not accessible anymore, it was replace by `luSelectOptions()` method which gives you the list of options
 ### Updated dependencies
 ### Updated devDependencies